+++ topic/topictypes/topictypes.go
@@ -3,6 +3,7 @@ package topictypes
 import (
 	"time"
 
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/clone"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/grpcwrapper/rawtopic"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/grpcwrapper/rawtopic/rawtopiccommon"
 )
@@ -27,6 +28,10 @@ const (
 	CodecCustomerFirst = Codec(rawtopiccommon.CodecCustomerFirst)
 	CodecCustomerEnd   = Codec(rawtopiccommon.CodecCustomerEnd) // last allowed custom codec id is CodecCustomerEnd-1
 )
+
+func (c Codec) ToRaw(r *rawtopiccommon.Codec) {
+	*r = rawtopiccommon.Codec(c)
+}
 
 // Consumer
 //
@@ -83,6 +88,37 @@ func (c *Consumer) FromRaw(raw *rawtopic.Consumer) {
 	}
 }
 
+// MeteringMode
+//
+// # Experimental
+//
+// Notice: This API is EXPERIMENTAL and may be changed or removed in a later release.
+type MeteringMode int
+
+const (
+	MeteringModeUnspecified      = MeteringMode(rawtopic.MeteringModeUnspecified)
+	MeteringModeReservedCapacity = MeteringMode(rawtopic.MeteringModeReservedCapacity)
+	MeteringModeRequestUnits     = MeteringMode(rawtopic.MeteringModeRequestUnits)
+)
+
+// FromRaw
+//
+// # Experimental
+//
+// Notice: This API is EXPERIMENTAL and may be changed or removed in a later release.
+func (m *MeteringMode) FromRaw(raw rawtopic.MeteringMode) {
+	*m = MeteringMode(raw)
+}
+
+// ToRaw
+//
+// # Experimental
+//
+// Notice: This API is EXPERIMENTAL and may be changed or removed in a later release.
+func (m *MeteringMode) ToRaw(raw *rawtopic.MeteringMode) {
+	*raw = rawtopic.MeteringMode(*m)
+}
+
 // PartitionSettings
 //
 // # Experimental
@@ -121,12 +157,15 @@ func (s *PartitionSettings) FromRaw(raw *rawtopic.PartitioningSettings) {
 type TopicDescription struct {
 	Path                              string
 	PartitionSettings                 PartitionSettings
-	Consumers                         []Consumer
-	SupportedCodecs                   []Codec
+	Partitions                        []PartitionInfo
 	RetentionPeriod                   time.Duration
+	RetentionStorageMB                int64
+	SupportedCodecs                   []Codec
 	PartitionWriteBurstBytes          int64
 	PartitionWriteSpeedBytesPerSecond int64
 	Attributes                        map[string]string
+	Consumers                         []Consumer
+	MeteringMode                      MeteringMode
 }
 
 // FromRaw
@@ -138,10 +177,13 @@ func (d *TopicDescription) FromRaw(raw *rawtopic.DescribeTopicResult) {
 	d.Path = raw.Self.Name
 	d.PartitionSettings.FromRaw(&raw.PartitioningSettings)
 
-	d.Consumers = make([]Consumer, len(raw.Consumers))
-	for i := 0; i < len(raw.Consumers); i++ {
-		d.Consumers[i].FromRaw(&raw.Consumers[i])
+	d.Partitions = make([]PartitionInfo, len(raw.Partitions))
+	for i := range raw.Partitions {
+		d.Partitions[i].FromRaw(&raw.Partitions[i])
 	}
+
+	d.RetentionPeriod = raw.RetentionPeriod
+	d.RetentionStorageMB = raw.RetentionStorageMB
 
 	d.SupportedCodecs = make([]Codec, len(raw.SupportedCodecs))
 	for i := 0; i < len(raw.SupportedCodecs); i++ {
@@ -155,4 +197,36 @@ func (d *TopicDescription) FromRaw(raw *rawtopic.DescribeTopicResult) {
 	for k, v := range raw.Attributes {
 		d.Attributes[k] = v
 	}
+
+	d.Consumers = make([]Consumer, len(raw.Consumers))
+	for i := 0; i < len(raw.Consumers); i++ {
+		d.Consumers[i].FromRaw(&raw.Consumers[i])
+	}
+
+	d.MeteringMode.FromRaw(raw.MeteringMode)
+}
+
+// PartitionInfo
+//
+// # Experimental
+//
+// Notice: This API is EXPERIMENTAL and may be changed or removed in a later release.
+type PartitionInfo struct {
+	PartitionID        int64
+	Active             bool
+	ChildPartitionIDs  []int64
+	ParentPartitionIDs []int64
+}
+
+// FromRaw
+//
+// # Experimental
+//
+// Notice: This API is EXPERIMENTAL and may be changed or removed in a later release.
+func (p *PartitionInfo) FromRaw(raw *rawtopic.PartitionInfo) {
+	p.PartitionID = raw.PartitionID
+	p.Active = raw.Active
+
+	p.ChildPartitionIDs = clone.Int64Slice(raw.ChildPartitionIDs)
+	p.ParentPartitionIDs = clone.Int64Slice(raw.ParentPartitionIDs)
 }